infoPage, components: ArtisProfile, HunterProfile, UploadModdal and social adjustments
