+++ frontend/src/App.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import LandingPage from './pages/landingPage'
+import Landing from './pages/landingPage'
 import Home from './pages/homePage';
 import Profile from './pages/profilePage';
 import Social from './pages/socialPage';
@@ -11,10 +11,11 @@ function App() {
   return (
     <Router>
       <Routes>
-        <Route path="/" element={<Home />} />
+        <Route path="/" element={<Profile />} />
         <Route path="/home" element={<Home />} />
         <Route path="/profile" element={<Profile />} />
         <Route path="/social" element={<Social />} />
+        <Route path="/landing" element={<Landing />} />
       </Routes>
     </Router>
   );
