+++ frontend/src/components/HunterProfile.sass
@@ -1,0 +1,120 @@
+#root, html, body
+  width: 100%
+  height: 100%
+  margin: 0
+  padding: 0
+
+.profile-container
+  width: 100%
+  height: 100%
+  background-color: #F5F5F5
+  display: flex
+  flex-direction: row
+  font-family: Amiko, sans-serif
+
+  .profile-items
+    width: 100%
+
+    .profile-page
+      display: flex
+      flex-direction: column
+      align-items: center
+      padding: 20px
+
+      h2
+        width: 100%
+        font-size: 24px
+        margin-bottom: 80px
+        align-self: flex-start
+        left: 60px
+        padding: 20px
+        border-bottom: 1px solid #ccc
+
+      .profile-form
+        display: flex
+        flex-direction: row
+
+        .profile-picture
+          display: flex
+          flex-direction: column
+          align-items: center
+          margin-bottom: 20px
+          margin-right: 100px
+
+          img
+            width: 200px
+            height: 200px
+            border-radius: 50%
+            margin-bottom: 10px
+
+          .upload-button
+            color: rgba(43, 39, 38, 0.78)
+            background: none
+            padding: 8px 12px
+            border: none
+            border-radius: 5px
+            cursor: pointer
+            
+        .form-group
+          display: flex
+          flex-direction: column
+          gap: 20px
+
+          .form-line
+            display: flex
+            gap: 20px
+
+            .form-field
+              display: flex
+              flex-direction: column
+              flex-grow: 1
+
+              label
+                font-size: 14px
+
+              input
+                padding: 10px
+                border: none
+                border-radius: 10px
+                font-size: 14px
+
+              .textarea
+                width: 100%
+                padding: 12px 20px
+                border: none
+                border-radius: 10px
+                background-color: #fff
+                font-size: 14px
+                resize: none
+                margin: 0
+
+            .email-field
+              display: flex
+              flex-direction: column
+              width: 49%
+
+              label
+                font-size: 14px
+
+              input
+                padding: 10px
+                border: none
+                border-radius: 10px
+                font-size: 14px
+
+          .password-field
+            display: flex
+            align-items: center
+
+            input
+              flex-grow: 1
+              width: 50%
+
+            .change-password
+              background-color: #e6d9ae
+              color: #2c2c2c
+              padding: 8px 12px
+              border: none
+              border-radius: 5px
+              margin-left: 10px
+              cursor: pointer
