+++ frontend/src/components/UploadModal.sass
@@ -71,6 +71,8 @@
 
         textarea
           height: 100px
+          resize: none
+          margin-bottom: 10px
 
     .modal-footer
       display: flex
