+++ frontend/src/pages/infoPage/info.sass
@@ -1,113 +1,210 @@
+#root, html, body
+  width: 100%
+  height: 100%
+  margin: 0
+  padding: 0
+
 .info-container
+  width: 100%
+  height: 100%
+  background-color: #F5F5F5
   display: flex
   flex-direction: row
-  background-color: #F9F9F9
-  height: 100vh
+  font-family: Amiko, sans-serif
 
-.sidebar
-  display: flex
-  justify-content: center
-  align-items: center
-  width: 15%
-  background-color: #FFFFFF
-  border-right: 1px solid #E0E0E0
+  .info-items
+    width: 100%
 
-.info-content
-  width: 85%
-  padding: 20px
-  background-color: #FFFFFF
-  border-radius: 10px
-  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1)
-  overflow-y: auto
+    .info-page
+      padding: 50px
 
-.header
-  display: flex
-  align-items: flex-start
-  margin-bottom: 20px
+    .back-button
+      background: none
+      border: none
+      color: #000000
+      font-size: 1.5em
+      margin-right: 20px
+      cursor: pointer
+      margin-top: -20px
+      margin-bottom: 10px
 
-  .back-button
-    background: none
-    border: none
-    color: #000000
-    font-size: 1.5em
-    margin-right: 20px
-    cursor: pointer
-    margin-top: -20px
-
-  .profile-section
-    display: flex
-    align-items: flex-start
-
-    .profile-pic
-      width: 48%
-      height: 160px
-      object-fit: cover
-      margin-right: 30px
-      border-radius: 100px
-
-    .profile-details
+    .profile-section
       display: flex
-      flex-direction: column
-      justify-content: flex-start
+      align-items: flex-start
 
-      .profile-name-section
+      .profile-picture
         display: flex
-        justify-content: space-between
+        flex-direction: column
         align-items: center
-
-        .profile-name
-          font-size: 2em
-          font-weight: bold
-
-        .profile-country
-          display: flex
-          align-items: center
-          margin-top: -10px
-
-          .flag-icon
-            font-size: 1.5em
-            margin-right: 5px
-
-      .profile-about-title
-        font-size: 1.2em
-        font-weight: bold
-        margin-top: 10px
-
-      .profile-about
-        margin-top: 10px
-        font-size: 1em
-        color: #666
-
-      .profile-info
-        display: flex
+        margin-bottom: 20px
         margin-top: 20px
+        margin-right: 100px
+        
+        img
+          width: 200px
+          height: 200px
+          border-radius: 50%
+          margin-bottom: 10px
 
-        .info-section
+      .profile-details
+        display: flex
+        flex-direction: column
+        justify-content: flex-start
+
+        .profile-name-section
           display: flex
-          flex-direction: column
+          justify-content: space-between
+          align-items: center
 
-          .info-title
+          .profile-name
+            font-size: 2em
             font-weight: bold
-            margin-bottom: 10px
-            font-size: 1.2em
 
-          .artistic-field, .languages-spoken
+          .profile-country
             display: flex
-            gap: 5px
-            margin-right: 40px
-          
-          .artistic-field
-            .tag
-              background-color: #E6D9AE
-              padding: 10px 20px
-              border-radius: 20px
-              color: #000000
-              font-size: 1em
+            align-items: center
+            margin-top: -10px
 
-          .languages-spoken
-            .tag
-              background-color: #2B2726
-              padding: 10px 20px
-              border-radius: 20px
-              color: #FFFFFF
-              font-size: 1em
+            .flag-icon
+              font-size: 1.5em
+              margin-right: 5px
+
+        .profile-about-title
+          font-size: 1.2em
+          font-weight: bold
+          margin-top: 10px
+
+        .profile-about
+          margin-top: 10px
+          font-size: 1em
+          color: #666
+
+        .profile-info
+          display: flex
+          margin-top: 20px
+
+          .info-section
+            display: flex
+            flex-direction: column
+
+            .info-title
+              font-weight: bold
+              margin-bottom: 10px
+              font-size: 1.2em
+
+            .artistic-field, .languages-spoken
+              display: flex
+              gap: 5px
+              margin-right: 40px
+            
+            .artistic-field
+              .tag
+                background-color: #E6D9AE
+                padding: 10px 20px
+                border-radius: 20px
+                color: #000000
+                font-size: 1em
+
+            .languages-spoken
+              .tag
+                background-color: #2B2726
+                padding: 10px 20px
+                border-radius: 20px
+                color: #FFFFFF
+                font-size: 1em
+  
+.my-works
+  margin-top: 40px
+  display: flex
+  flex-direction: column
+  align-items: center
+  justify-content: center
+
+  h3
+    font-size: 24px
+    margin-bottom: 20px
+    align-self: center
+
+  .portfolio-tabs
+    display: flex
+    gap: 20px
+    margin-bottom: 20px
+    align-items: center
+    justify-content: center
+
+    .tab
+      background-color: #f5f5f5
+      color: #2c2c2c
+      padding: 10px 20px
+      border: none
+      border-radius: 5px
+      cursor: pointer
+      font-size: 16px
+      &.active
+        background-color: #e6d9ae
+        color: #2c2c2c
+
+  .portfolio-content
+    display: flex
+    flex-direction: row
+    gap: 20px
+    flex-wrap: wrap
+    align-items: center
+    justify-content: center
+
+    .audio-item
+      display: flex
+      flex-direction: row
+      padding: 15px
+      background-color: #fff
+      border-radius: 10px
+      box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1)
+      width: 400px
+
+      .audio-icon
+        font-size: 24px
+        margin-bottom: 10px
+
+      .audio-details
+        display: flex
+        flex-direction: column
+        align-items: flex-start
+        justify-content: flex-start
+
+        p
+          margin: 2px 18px
+          font-size: 14px
+          color: #2c2c2c
+
+.send-message
+  margin-top: 40px
+
+  h4
+    font-size: 20px
+    margin-bottom: 10px
+
+  .message-input-container
+    display: flex
+    align-items: center
+    background-color: #fff
+    border-radius: 10px
+    padding: 10px
+    box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1)
+
+    input
+      flex: 1
+      border: none
+      outline: none
+      font-size: 14px
+      padding: 10px
+
+    .send-button
+      background-color: transparent
+      border: none
+      font-size: 24px
+      cursor: pointer
+      padding: 10px
+      
+    .send-icon
+      cursor: pointer
+      color: #9B7243
