+++ frontend/src/pages/socialPage/social.sass
@@ -10,6 +10,7 @@
   height: 100%
   display: flex
   flex-direction: row
+  font-family: Amiko, sans-serif
 
 .chat-history
   width: 30%
